Adds support for getText style translations
The fallback language is set to "false" to use the keys
as the translation.
Additionally because the en.pot file will not have any translations
as the keys are already english, we won't allow emptyString values
to be used as translations. Those translations will therefore fall
back to the "key".